fix out of bounds error
to build with Cosmopolitan's ASAN/UBSAN runtime, uncomment the line with
MODE = dbg in the Makefile.

after building with Cosmopolitan's UBSAN runtime, we find out that there
is an out of bounds error that occurs if the file we read has 16
characters or more. now we have a fix for that.